+++ src/gzm/ontology/clustering/Clustering.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public abstract class Clustering {
-	abstract void buildClusters(ArrayList<List<Float>> Sim);
+	abstract void buildClusters(ArrayList<List<Double>> Sim);
 	abstract ArrayList<Cluster> getClusters();
 	
 }
